[evotsc] Add neutral evolution mode
This commit introduces a neutral evolution mode, intended to measure
various genomic parameters under a neutral model (no selection).

In order to toggle this mode, a new parameter --neutral is introduced to
the main executable `evotsc_run.py`.

+++ evotsc_run.py
@@ -1,5 +1,6 @@
 import sys
 import os
+import argparse
 import pickle
 import pathlib
 import re
@@ -70,14 +71,22 @@ def write_stats(stats_file, indiv, gen):
                     f'{on_genes_B[0]},{off_genes_B[0]},{on_genes_B[1]},{off_genes_B[1]},{on_genes_B[2]},{off_genes_B[2]}\n')
     stats_file.flush()
 
-def main():
-    # Parse CLI arguments: output directory and number of generations
-    if len(sys.argv) != 3:
-        print('Usage: `run_evotsc output_dir nb_generations`. Exiting.')
-        return
 
-    output_dir = pathlib.Path(sys.argv[1])
-    nb_generations = int(sys.argv[2])
+def main():
+
+    # Parse CLI arguments
+    arg_parser = argparse.ArgumentParser()
+    arg_parser.add_argument('-n', '--generations', type=int, required=True,
+                            help='number of generations')
+    arg_parser.add_argument('-o', '--output_dir', required=True,
+                            help='output directory')
+    arg_parser.add_argument('--neutral', action='store_true',
+                            help='run without selection')
+    args = arg_parser.parse_args()
+
+    nb_generations = int(args.generations)
+    output_dir = pathlib.Path(args.output_dir)
+
 
     first_start = True
     start_gen = 1
@@ -117,12 +126,13 @@ def main():
                                        sigma_A=sigma_A,
                                        sigma_B=sigma_B)
 
-        stats_file = open(f'{output_dir}/stats.csv', 'w')
-        stats_file.write('Gen,Fitness,ABon_A,ABoff_A,Aon_A,Aoff_A,Bon_A,Boff_A,'
-                                    'ABon_B,ABoff_B,Aon_B,Aoff_B,Bon_B,Boff_B\n')
+        if not args.neutral:
+            stats_file = open(f'{output_dir}/stats.csv', 'w')
+            stats_file.write('Gen,Fitness,ABon_A,ABoff_A,Aon_A,Aoff_A,Bon_A,Boff_A,'
+                                        'ABon_B,ABoff_B,Aon_B,Aoff_B,Bon_B,Boff_B\n')
 
-        save_indiv(output_dir, init_indiv, 0)
-        write_stats(stats_file, init_indiv, 0)
+            save_indiv(output_dir, init_indiv, 0)
+            write_stats(stats_file, init_indiv, 0)
 
     else:
         save_files = [f for f in output_dir.iterdir() if 'pop_gen' in f.name]
@@ -130,32 +140,42 @@ def main():
         start_gen = int(re.search(r'\d+', last_save_path.name).group(0)) + 1
         population = load_pop(last_save_path)
 
-        # Get rid of the stats that happened between the last save and the crash
-        os.rename(f'{output_dir}/stats.csv', f'{output_dir}/old_stats.csv')
+        if not args.neutral:
+            # Get rid of the stats that happened between the last save and the crash
+            os.rename(f'{output_dir}/stats.csv', f'{output_dir}/old_stats.csv')
 
-        old_stats_path = f'{output_dir}/old_stats.csv'
-        old_stats_file = open(old_stats_path)
-        stats_file = open(f'{output_dir}/stats.csv', 'w')
-        for gen in range(start_gen+1):
-            stats_file.write(old_stats_file.readline())
+            old_stats_path = f'{output_dir}/old_stats.csv'
+            old_stats_file = open(old_stats_path)
+            stats_file = open(f'{output_dir}/stats.csv', 'w')
+            for gen in range(start_gen+1):
+                stats_file.write(old_stats_file.readline())
 
-        old_stats_file.close()
-        os.remove(old_stats_path)
+            old_stats_file.close()
+            os.remove(old_stats_path)
 
 
-    for gen in range(start_gen, nb_generations+1):
-        best_indiv, avg_fit = population.step()
+    if not args.neutral:
+        for gen in range(start_gen, nb_generations+1):
+            best_indiv, avg_fit = population.step()
 
-        print(f'Gen {gen}: best fit {best_indiv.fitness:.5}, avg fit {avg_fit:.5}')
-        write_stats(stats_file, best_indiv, gen)
+            print(f'Gen {gen}: best fit {best_indiv.fitness:.5}, avg fit {avg_fit:.5}')
+            write_stats(stats_file, best_indiv, gen)
 
-        if gen % save_best_step == 0:
-            save_indiv(output_dir, best_indiv, gen)
+            if gen % save_best_step == 0:
+                save_indiv(output_dir, best_indiv, gen)
 
-        if gen % save_full_step == 0:
-            save_pop(output_dir, population, gen)
+            if gen % save_full_step == 0:
+                save_pop(output_dir, population, gen)
 
-    stats_file.close()
+        stats_file.close()
+
+    else: # Neutral evolution -- no selection
+        for gen in range(start_gen, nb_generations+1):
+            population.neutral_step()
+            print(f'Generation {gen}')
+
+            if gen % save_full_step == 0:
+                save_pop(output_dir, population, gen)
 
 
 if __name__ == "__main__":
